+++ R/svyreg.R
@@ -2,18 +2,22 @@
 svyreg <- function(formula, design, var = NULL, na.rm = FALSE)
 {
     dat <- .check_regression(formula, design, var, NULL, na.rm)
-    # account for heteroscedasticity
-    if (!is.null(var)) {
-        dat$x <- dat$x / sqrt(dat$var)
-        dat$y <- dat$y / sqrt(dat$var)
-    }
     n <- length(dat$y); p <- NCOL(dat$x)
+    # account for heteroscedasticity
+    if (is.null(var)) {
+        x <- dat$x
+        y <- dat$y
+    } else {
+        x <- dat$x / sqrt(dat$var)
+        y <- dat$y / sqrt(dat$var)
+    }
     if (n < p)
         stop("Number of observations cannot be smaller than no. of variables\n")
-    tmp <- .C("wlslm", x = as.double(dat$x), y = as.double(dat$y),
+    tmp <- .C("wlslm", x = as.double(x), y = as.double(y),
         w = as.double(dat$w), resid = as.double(numeric(n)),
         n = as.integer(n), p = as.integer(p), beta = as.double(numeric(p)),
         scale = as.double(numeric(1)), PACKAGE = "robsurvey")
+    # Note: The residuals are (y - x*beta) / sqrt(var)
     names(tmp$beta) <- colnames(dat$x)
     # return
     structure(list(characteristic = "regression",